DataPlatformEtlService/requirements.prod.txt
setuptools==80.7.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
PyJWT==2.10.1
pandas==2.2.3
numpy==2.2.6
fastapi==0.115.12
boto3==1.38.19
google-cloud-storage==3.1.0
google-cloud-pubsub==2.29.0
huggingface-hub==0.31.4
pymongo==4.13.0
PyYAML==6.0.2
clickhouse-connect==0.8.17
pyspark==3.5.5
pyodbc==5.2.0
pyarrow==20.0.0
sqlparse==0.5.3
duckdb==1.2.2
sqlparse==0.5.3
Jinja2==3.1.6
